flask-servicio2/requirements.txt
flask==2.3.2
ariadne==0.20.1
requests==2.31.0
python-dotenv==1.0.0